+++ packages/app/cypress/integration/ArgoCD.ts
@@ -22,7 +22,7 @@ describe('Travis CI', () => {
     cy.saveGithubToken();
     cy.intercept(
       'GET',
-      'http://localhost:7007/api/proxy/argocd/api/applications/name/test-app',
+      'http://localhost:7007/api/proxy/argocd/api/applications/test-app',
       { fixture: 'ArgoCD/applications-test-app.json' },
     );
     cy.visit('/catalog/default/component/sample-service');