+++ plugins/frontend/backstage-plugin-aws-lambda/src/hooks/useLambda.ts
@@ -13,8 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { useAsyncRetry } from 'react-use';
-import { useApi, errorApiRef, configApiRef, identityApiRef } from '@backstage/core-plugin-api';
+import { useAsync, useAsyncRetry } from 'react-use';
+import {
+  useApi,
+  errorApiRef,
+  configApiRef,
+  identityApiRef,
+} from '@backstage/core-plugin-api';
 import { LambdaData } from '../types';
 import { awsLambdaApiRef } from '../api';
 import { useCallback } from 'react';
@@ -30,14 +35,20 @@ export function useLambda({
   const errorApi = useApi(errorApiRef);
   const configApi = useApi(configApiRef);
   const identityApi = useApi(identityApiRef);
+
+  const identity = useAsync(async () => {
+    return await identityApi.getCredentials();
+  });
+
   const getFunctionByName = useCallback(
-    async () =>
-      lambdaApi.getFunctionByName({
+    async () => {
+      return await lambdaApi.getFunctionByName({
         backendUrl: configApi.getString('backend.baseUrl'),
         awsRegion: region,
         functionName: lambdaName,
-        token: identityApi?.getIdToken ? await identityApi.getIdToken() : undefined
-      }),
+        token: identity.value?.token || undefined,
+      });
+    },
     [region, lambdaName], // eslint-disable-line react-hooks/exhaustive-deps
   );
   const {
@@ -49,7 +60,7 @@ export function useLambda({
     try {
       const lambdaFunction = await getFunctionByName();
       return lambdaFunction;
-    } catch (err:any) {
+    } catch (err) {
       if (err?.message === 'MissingBackendAwsAuthException') {
         errorApi.post(new Error('Please add aws auth backend plugin'));
         return null;