+++ plugins/frontend/backstage-plugin-bugsnag/src/api/BugsnagClient.ts
@@ -49,9 +49,18 @@ export class BugsnagClient implements BugsnagApi {
     return payload;
   }
 
-  async fetchProjects(organisationId: string): Promise<Project[]> {
+  async fetchProjects({
+    organisationId,
+    projectName,
+    perPage = 30,
+  }: {
+    organisationId?: string;
+    projectName?: string;
+    perPage?: number;
+  }): Promise<Project[]> {
+    const query = projectName ? `q=${projectName}&` : '';
     const response = await fetch(
-      `${await this.getApiUrl()}/organizations/${organisationId}/projects`,
+      `${await this.getApiUrl()}/organizations/${organisationId}/projects?${query}per_page=${perPage}`,
     );
     const payload = await response.json();
     if (!response.ok) {
@@ -60,9 +69,15 @@ export class BugsnagClient implements BugsnagApi {
     return await payload;
   }
 
-  async fetchErrors(projectId: string): Promise<BugsnagError[]> {
+  async fetchErrors({
+    projectId,
+    perPage = 30,
+  }: {
+    projectId: string;
+    perPage?: number;
+  }): Promise<BugsnagError[]> {
     const response = await fetch(
-      `${await this.getApiUrl()}/projects/${projectId}/errors`,
+      `${await this.getApiUrl()}/projects/${projectId}/errors?per_page=${perPage}`,
     );
     const payload = await response.json();
     if (!response.ok) {