+++ plugins/frontend/backstage-plugin-github-insights/src/components/InsightsPage/InsightsPage.test.tsx
@@ -19,26 +19,25 @@ import { act, render, RenderResult, waitFor } from '@testing-library/react';
 import { InsightsPage } from './InsightsPage';
 import { ThemeProvider } from '@material-ui/core';
 import { lightTheme } from '@backstage/theme';
-import { wrapInTestApp } from '@backstage/test-utils';
+import { wrapInTestApp, TestApiProvider } from '@backstage/test-utils';
 import { EntityProvider } from '@backstage/plugin-catalog-react';
 
-import { githubAuthApiRef } from '@backstage/core-plugin-api';
+import { AnyApiRef, githubAuthApiRef } from '@backstage/core-plugin-api';
 
 import {
-  ApiProvider,
-  ApiRegistry,
   GithubAuth,
   OAuthRequestManager,
   UrlPatternDiscovery,
+  ConfigReader,
 } from '@backstage/core-app-api';
-import { scmIntegrationsApiRef } from '@backstage/integration-react';
 import {
-  createScmIntegrationsApiMock,
-  defaultIntegrationsConfig,
-} from '../../mocks/scmIntegrationsApiMock';
+  scmIntegrationsApiRef,
+  ScmIntegrationsApi,
+} from '@backstage/integration-react';
+import { defaultIntegrationsConfig } from '../../mocks/scmIntegrationsApiMock';
 
 const oauthRequestApi = new OAuthRequestManager();
-const apis = ApiRegistry.from([
+const apis: [AnyApiRef, Partial<unknown>][] = [
   [
     githubAuthApiRef,
     GithubAuth.create({
@@ -50,9 +49,16 @@ const apis = ApiRegistry.from([
   ],
   [
     scmIntegrationsApiRef,
-    createScmIntegrationsApiMock(defaultIntegrationsConfig),
+    ScmIntegrationsApi.fromConfig(
+      ConfigReader.fromConfigs([
+        {
+          context: 'unit-test',
+          data: defaultIntegrationsConfig,
+        },
+      ]),
+    ),
   ],
-]);
+];
 
 describe('Insights Page', () => {
   it('should render', async () => {
@@ -61,7 +67,7 @@ describe('Insights Page', () => {
     await act(async () => {
       renderResult = render(
         wrapInTestApp(
-          <ApiProvider apis={apis}>
+          <TestApiProvider apis={apis}>
             <ThemeProvider theme={lightTheme}>
               <EntityProvider
                 entity={{
@@ -80,7 +86,7 @@ describe('Insights Page', () => {
                 <InsightsPage />
               </EntityProvider>
             </ThemeProvider>
-          </ApiProvider>,
+          </TestApiProvider>,
         ),
       );
     });