+++ utils/roadie-backstage-entity-validator/src/validator.js
@@ -12,12 +12,17 @@ import {
   locationEntityV1alpha1Validator,
   templateEntityV1beta2Validator,
   userEntityV1alpha1Validator,
+  systemEntityV1alpha1Validator,
+  domainEntityV1alpha1Validator,
+  resourceEntityV1alpha1Validator
 } from '@backstage/catalog-model';
 import annotationSchema from './schemas/annotations.schema.json';
 import Ajv from 'ajv';
+import ajvFormats from 'ajv-formats'
+import path from 'path'
 
 const ajv = new Ajv({ verbose: true });
-require('ajv-formats')(ajv);
+ajvFormats(ajv);
 
 const VALIDATORS = {
   api: apiEntityV1alpha1Validator,
@@ -26,6 +31,9 @@ const VALIDATORS = {
   location: locationEntityV1alpha1Validator,
   template: templateEntityV1beta2Validator,
   user: userEntityV1alpha1Validator,
+  system: systemEntityV1alpha1Validator,
+  domain: domainEntityV1alpha1Validator,
+  resource: resourceEntityV1alpha1Validator
 };
 
 function modifyPlaceholders(obj) {
@@ -115,6 +123,49 @@ export const validate = async (fileContents, verbose = true) => {
   }
 };
 
+const relativeSpaceValidation = async (fileContents, filepath, verbose) => {
+  const fileExists = (fp) => {
+    let flag = true;
+    try{
+      fs.accessSync(fp, fs.constants.F_OK);
+    } catch(e){
+      flag = false;
+    }
+    return flag;
+  }
+
+  const validateTechDocs = async (data, fp) => {
+    if(!data?.metadata?.annotations || !data?.metadata?.annotations["backstage.io/techdocs-ref"] ){
+      return
+    }
+    const techDocsAnnotation = data.metadata.annotations["backstage.io/techdocs-ref"]
+    if(!techDocsAnnotation.includes('dir')){
+      return
+    }
+
+    const mkdocsPath = path.join(path.dirname(fp), techDocsAnnotation.split(':')[1], 'mkdocs.yaml');
+
+    if(await !fileExists(mkdocsPath)){
+      throw new Error(`Techdocs annotation specifies "dir" but file under ${mkdocsPath} not found`)
+    }
+    return
+  }
+
+  try {
+    const data = yaml.loadAll(fileContents);
+    if(verbose){
+      console.log('Validating locally dependant catalog contents')
+    }
+    await Promise.all(
+      data.map(async (it) => {
+        await validateTechDocs(it, filepath);
+      })
+    );
+  } catch(e){
+    throw new Error(e);
+  }
+}
+
 export const validateFromFile = async (
   filepath = './sample/catalog-info.yml',
   verbose = true,
@@ -124,4 +175,5 @@ export const validateFromFile = async (
     console.log(`Validating Entity Schema policies for file ${filepath}`);
   }
   await validate(fileContents, verbose);
+  await relativeSpaceValidation(fileContents, filepath, verbose)
 };