+++ plugins/frontend/backstage-plugin-travis-ci/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @roadiehq/backstage-plugin-travis-ci
 
+## 1.3.6
+
+### Patch Changes
+
+- 5ae1b4b: Update README.md files
+
+## 1.3.5
+
+### Patch Changes
+
+- b5db653: Update dependecies to latest packages
+
+## 1.3.4
+
+### Patch Changes
+
+- a02dbf5: Migrating to TestApiProvider in the tests
+
+## 1.3.3
+
+### Patch Changes
+
+- 142ce1c: Moved React dependencies to `peerDependencies` and allow both React v16 and v17 to be used.
+- faf6526: Removed unused dependency on `@backstage/core-app-api`.
+
+## 1.3.2
+
+### Patch Changes
+
+- 49abec7: Update patch to release new changes.
+
 ## 1.3.1
 
 ### Patch Changes